// src/auth/enums/permission.enum.ts
export enum Permission {
    POST_CREATE = 'POST_CREATE',

    POST_UPDATE_OWN = 'POST_UPDATE_OWN',
    POST_UPDATE_ANY = 'POST_UPDATE_ANY',
    POST_DELETE_OWN = 'POST_DELETE_OWN',
    POST_DELETE_ANY = 'POST_DELETE_ANY',

    COMMENT_CREATE = 'COMMENT_CREATE',
    COMMENT_DELETE_OWN = 'COMMENT_DELETE_OWN',
    COMMENT_DELETE_ANY = 'COMMENT_DELETE_ANY',
    COMMENT_UPDATE_OWN = 'COMMENT_UPDATE_OWN',

    REACTION_CREATE = 'REACTION_CREATE',
    REACTION_DELETE_OWN = 'REACTION_DELETE_OWN',

    USER_LIST = 'USER_LIST',
    USER_CHANGE_ROLE = 'USER_CHANGE_ROLE',
    USER_VIEW_DETAILS = 'USER_VIEW_DETAILS',

    ROLE_PERMISSION_VIEW = 'ROLE_PERMISSION_VIEW',
    ROLE_PERMISSION_UPDATE = 'ROLE_PERMISSION_UPDATE',

    EXTERNAL_USERS_VIEW = 'EXTERNAL_USERS_VIEW',
    EXTERNAL_POST_CREATE = 'EXTERNAL_POST_CREATE',
}